Fix : Entity 칼럼 일부 수정(isDelete, 조인 네이밍) (#9)
