+++ src/main/java/com/hermes/ithermes/domain/entity/ContentsViewMapping.java
@@ -11,7 +11,7 @@ import lombok.RequiredArgsConstructor;
 @Builder
 @RequiredArgsConstructor
 @AllArgsConstructor
-public class ContentsViewMapping {
+public class ContentsViewMapping extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
