+++ src/main/java/com/hermes/ithermes/domain/entity/Job.java
@@ -2,13 +2,13 @@ package com.hermes.ithermes.domain.entity;
 
 import jakarta.persistence.*;
 
-public class Job {
+public class Job extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "SERVICE_ID")
+    @JoinColumn(name = "serviceId")
     Service service;
 
     @Column(nullable = false)
@@ -31,6 +31,9 @@ public class Job {
 
     @Column(nullable = false)
     private Long viewCount;
+
+    @Column(nullable = false)
+    private Boolean isDelete;
 
     public void setService(Service service) {
         this.service = service;
