+++ src/main/java/com/hermes/ithermes/domain/entity/YoutubeAndNews.java
@@ -17,7 +17,7 @@ public class YoutubeAndNews extends BaseEntity {
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "SERVICE_ID")
+    @JoinColumn(name = "serviceId")
     Service service;
 
     @Column(nullable = false)
@@ -37,6 +37,9 @@ public class YoutubeAndNews extends BaseEntity {
 
     @Column(nullable = false)
     private Long viewCount;
+
+    @Column(nullable = false)
+    private Boolean isDelete;
 
     public void setService(Service service) {
         this.service = service;
